+++ scrapy/linkextractors/__init__.py
@@ -1,6 +1,9 @@
 """
-Common code and definitions used by Link extractors (located in
-scrapy.linkextractors).
+scrapy.linkextractors
+
+This package contains a collection of Link Extractors.
+
+For more info see docs/topics/link-extractors.rst
 """
 import re
 from six.moves.urllib.parse import urlparse
@@ -98,3 +101,6 @@ class FilteringLinkExtractor(object):
 
     def _extract_links(self, *args, **kwargs):
         return self.link_extractor._extract_links(*args, **kwargs)
+
+# Top-level imports
+from .lxmlhtml import LxmlLinkExtractor as LinkExtractor
